Restructuring Hook logic inside the Hook chain. (#83)
