+++ Makefile
@@ -4,9 +4,17 @@ all: build upload
 
 build:
 	mkdir -p build
-	wasmcc ./src/evernode.c -o ./build/evernode.wasm -O0 -Wl,--allow-undefined -I../
-	wasm-opt -O2 ./build/evernode.wasm -o ./build/evernode.wasm
-	hook-cleaner ./build/evernode.wasm
+	wasmcc ./src/evernodezero.c -o ./build/evernodezero.wasm -O0 -Wl,--allow-undefined -I../
+	wasm-opt -O2 ./build/evernodezero.wasm -o ./build/evernodezero.wasm
+	hook-cleaner ./build/evernodezero.wasm
+
+	wasmcc ./src/evernodeone.c -o ./build/evernodeone.wasm -O0 -Wl,--allow-undefined -I../
+	wasm-opt -O2 ./build/evernodeone.wasm -o ./build/evernodeone.wasm
+	hook-cleaner ./build/evernodeone.wasm
+
+	wasmcc ./src/evernodetwo.c -o ./build/evernodetwo.wasm -O0 -Wl,--allow-undefined -I../
+	wasm-opt -O2 ./build/evernodetwo.wasm -o ./build/evernodetwo.wasm
+	hook-cleaner ./build/evernodetwo.wasm
 
 upload:
 	node sethook.js
