+++ src/constants.h
@@ -132,4 +132,25 @@ const uint32_t MOMENT_BASE_POINT_OFFSET = 0;
 const uint32_t MOMENT_AT_TRANSITION_OFFSET = 8;
 const uint32_t MOMENT_TYPE_OFFSET = 12;
 
+#define COMMON_CHAIN_PARAMS "common_params"
+const uint32_t COMMON_CHAIN_PARAMS_SIZE = 57;
+const uint32_t CHAIN_IDX_PARAM_OFFSET = 0;
+const uint32_t CUR_LEDGER_SEQ_PARAM_OFFSET = 1;
+const uint32_t CUR_LEDGER_TIMESTAMP_PARAM_OFFSET = 9;
+const uint32_t HOOK_ACCID_PARAM_OFFSET = 17;
+const uint32_t ACCOUNT_FIELD_PARAM_OFFSET = 37;
+
+#define CHAIN_ONE_PARAMS "chain_one_params"
+const uint32_t CHAIN_ONE_PARAMS_SIZE = 88;
+const uint32_t AMOUNT_BUF_PARAM_OFFSET = 0;
+const uint32_t FLOAT_AMT_PARAM_OFFSET = 48;
+const uint32_t TXID_PARAM_OFFSET = 56;
+
+#define CHAIN_TWO_PARAMS "chain_two_params"
+const uint32_t CHAIN_TWO_PARAMS_SIZE = 28;
+const uint32_t MOMENT_BASE_IDX_PARAM_OFFSET = 0;
+const uint32_t CUR_MOMENT_TYPE_PARAM_OFFSET = 8;
+const uint32_t CUR_IDX_PARAM_OFFSET = 9;
+const uint32_t MOMENT_TRANSITION_INFO_PARAM_OFFSET = 17;
+
 #endif
